pyproject.toml
[tool.poetry]
name = "vkus"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
sqlalchemy = "^2.0.4"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
flask-sqlalchemy = "^3.0.3"
beautifulsoup4 = "^4.11.2"
flask-wtf = "^1.1.1"
pytest = "^7.2.2"
flask-migrate = "^4.0.4"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
jupyter = "^1.0.0"
notebook = "^6.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
